feat(rust-analyzer-plugin): add comprehensive argument validation and dispatch layer tests
- Introduce extensive unit tests for argument validation in the rust-analyzer plugin
- Add stdin/stdout dispatch layer tests to ensure robustness of plugin request handling
- Refactor test modules to separate support, argument validation, and dispatch tests
- Enhance URI handling in rename-symbol mapping with error propagation
- Update documentation with markdownlint and formatting improvements

This change improves reliability by validating plugin request arguments and verifying dispatch behavior under various input scenarios.
