+++ crates/weaver-plugin-rust-analyzer/src/tests/mod.rs
@@ -1,189 +1,28 @@
 //! Unit and behavioural tests for the rust-analyzer actuator plugin.
 
+mod argument_validation;
 mod behaviour;
+mod dispatch_layer;
+mod support;
 
-use std::collections::HashMap;
-use std::path::PathBuf;
-
-use mockall::mock;
-use rstest::{fixture, rstest};
+use rstest::rstest;
 use weaver_plugins::capability::ReasonCode;
-use weaver_plugins::protocol::{
-    DiagnosticSeverity, FilePayload, PluginOutput, PluginRequest, PluginResponse,
+use weaver_plugins::protocol::{PluginOutput, PluginRequest};
+
+use crate::{RustAnalyzerAdapterError, execute_request};
+use support::{
+    adapter_returning, adapter_returning_with_path, adapter_unused, rename_arguments,
+    request_with_args, request_with_path,
 };
 
-use crate::{
-    ByteOffset, RustAnalyzerAdapter, RustAnalyzerAdapterError, execute_request, run_with_adapter,
-};
-
-mock! {
-    Adapter {}
-    impl RustAnalyzerAdapter for Adapter {
-        fn rename(
-            &self,
-            file: &FilePayload,
-            offset: ByteOffset,
-            new_name: &str,
-        ) -> Result<String, RustAnalyzerAdapterError>;
-    }
-}
-
-/// Builds a `MockAdapter` that expects a single rename call returning `result`.
-fn adapter_returning(result: Result<String, RustAnalyzerAdapterError>) -> MockAdapter {
-    adapter_returning_with_path(result, None)
-}
-
-/// Builds a `MockAdapter` that can also assert the forwarded payload path.
-fn adapter_returning_with_path(
-    result: Result<String, RustAnalyzerAdapterError>,
-    expected_payload_path: Option<&str>,
-) -> MockAdapter {
-    let expected_path_string = expected_payload_path.map(String::from);
-    let mut adapter = MockAdapter::new();
-    adapter
-        .expect_rename()
-        .once()
-        .return_once(move |file, offset, new_name| {
-            if let Some(path) = &expected_path_string {
-                assert_eq!(file.path(), PathBuf::from(path).as_path());
-            }
-            assert_eq!(offset, ByteOffset::new(3));
-            assert_eq!(new_name, "new_name");
-            result
-        });
-    adapter
-}
-
-/// Builds a `MockAdapter` where rename is never expected.
-fn adapter_unused() -> MockAdapter {
-    MockAdapter::new()
-}
-
-#[fixture]
-fn rename_arguments() -> HashMap<String, serde_json::Value> {
-    let mut arguments = HashMap::new();
-    arguments.insert(
-        String::from("uri"),
-        serde_json::Value::String(String::from("file:///src/main.rs")),
-    );
-    arguments.insert(
-        String::from("position"),
-        serde_json::Value::String(String::from("3")),
-    );
-    arguments.insert(
-        String::from("new_name"),
-        serde_json::Value::String(String::from("new_name")),
-    );
-    arguments
-}
-
-fn request_with_args(arguments: HashMap<String, serde_json::Value>) -> PluginRequest {
-    PluginRequest::with_arguments(
-        "rename-symbol",
-        vec![FilePayload::new(
-            PathBuf::from("src/main.rs"),
-            "fn old_name() -> i32 {\n    1\n}\n",
-        )],
-        arguments,
-    )
-}
-
-#[rstest]
-fn rename_success_returns_diff_output(rename_arguments: HashMap<String, serde_json::Value>) {
+#[test]
+fn rename_success_returns_diff_output() {
     let adapter = adapter_returning(Ok(String::from("fn new_name() -> i32 {\n    1\n}\n")));
 
-    let response = execute_request(&adapter, &request_with_args(rename_arguments))
+    let response = execute_request(&adapter, &request_with_args(rename_arguments()))
         .expect("execute_request should succeed");
     assert!(response.is_success());
     assert!(matches!(response.output(), PluginOutput::Diff { .. }));
-}
-
-fn remove_uri(arguments: &mut HashMap<String, serde_json::Value>) {
-    arguments.remove("uri");
-}
-
-fn set_empty_uri(arguments: &mut HashMap<String, serde_json::Value>) {
-    arguments.insert(
-        String::from("uri"),
-        serde_json::Value::String(String::from("  ")),
-    );
-}
-
-fn set_numeric_uri(arguments: &mut HashMap<String, serde_json::Value>) {
-    arguments.insert(
-        String::from("uri"),
-        serde_json::Value::Number(serde_json::Number::from(4)),
-    );
-}
-
-fn remove_position(arguments: &mut HashMap<String, serde_json::Value>) {
-    arguments.remove("position");
-}
-
-fn set_boolean_position(arguments: &mut HashMap<String, serde_json::Value>) {
-    arguments.insert(String::from("position"), serde_json::Value::Bool(true));
-}
-
-fn set_negative_position(arguments: &mut HashMap<String, serde_json::Value>) {
-    arguments.insert(
-        String::from("position"),
-        serde_json::Value::String(String::from("-1")),
-    );
-}
-
-fn set_numeric_position(arguments: &mut HashMap<String, serde_json::Value>) {
-    arguments.insert(
-        String::from("position"),
-        serde_json::Value::Number(serde_json::Number::from(3)),
-    );
-}
-
-fn set_empty_new_name(arguments: &mut HashMap<String, serde_json::Value>) {
-    arguments.insert(
-        String::from("new_name"),
-        serde_json::Value::String(String::from("  ")),
-    );
-}
-
-fn set_numeric_new_name(arguments: &mut HashMap<String, serde_json::Value>) {
-    arguments.insert(
-        String::from("new_name"),
-        serde_json::Value::Number(serde_json::Number::from(42)),
-    );
-}
-
-#[rstest]
-#[case::missing_uri(remove_uri as fn(&mut _), Some("uri"))]
-#[case::empty_uri(set_empty_uri as fn(&mut _), Some("uri"))]
-#[case::numeric_uri(set_numeric_uri as fn(&mut _), Some("uri argument must be a string"))]
-#[case::missing_position(remove_position as fn(&mut _), Some("position"))]
-#[case::boolean_position(set_boolean_position as fn(&mut _), Some("position"))]
-#[case::negative_position(set_negative_position as fn(&mut _), Some("non-negative integer"))]
-#[case::numeric_position_succeeds(set_numeric_position as fn(&mut _), None)]
-#[case::numeric_new_name(set_numeric_new_name as fn(&mut _), Some("new_name argument must be a string"))]
-#[case::empty_new_name(set_empty_new_name as fn(&mut _), Some("new_name"))]
-fn rename_argument_validation(
-    #[case] mutate: fn(&mut HashMap<String, serde_json::Value>),
-    #[case] expected_error: Option<&str>,
-    mut rename_arguments: HashMap<String, serde_json::Value>,
-) {
-    mutate(&mut rename_arguments);
-
-    if let Some(needle) = expected_error {
-        let adapter = adapter_unused();
-        let err = execute_request(&adapter, &request_with_args(rename_arguments))
-            .expect_err("invalid arguments should fail");
-        assert!(
-            err.message().contains(needle),
-            "expected error mentioning '{needle}', got: {err}"
-        );
-        assert_eq!(err.reason_code(), Some(ReasonCode::IncompletePayload));
-    } else {
-        let adapter = adapter_returning(Ok(String::from("fn new_name() -> i32 {\n    1\n}\n")));
-        let response = execute_request(&adapter, &request_with_args(rename_arguments))
-            .expect("valid arguments should succeed");
-        assert!(response.is_success());
-    }
 }
 
 #[test]
@@ -213,24 +52,22 @@ enum FailureScenario {
 #[case::uri_mismatch(FailureScenario::UriMismatch)]
 #[case::relative_uri(FailureScenario::RelativeUri)]
 #[case::invalid_uri(FailureScenario::InvalidUri)]
-fn rename_non_mutating_or_error_returns_failure(
-    #[case] scenario: FailureScenario,
-    mut rename_arguments: HashMap<String, serde_json::Value>,
-) {
+fn rename_non_mutating_or_error_returns_failure(#[case] scenario: FailureScenario) {
+    let mut arguments = rename_arguments();
     if matches!(scenario, FailureScenario::UriMismatch) {
-        rename_arguments.insert(
+        arguments.insert(
             String::from("uri"),
             serde_json::Value::String(String::from("file:///src/other.rs")),
         );
     }
     if matches!(scenario, FailureScenario::RelativeUri) {
-        rename_arguments.insert(
+        arguments.insert(
             String::from("uri"),
             serde_json::Value::String(String::from("file:///./src/main.rs")),
         );
     }
     if matches!(scenario, FailureScenario::InvalidUri) {
-        rename_arguments.insert(
+        arguments.insert(
             String::from("uri"),
             serde_json::Value::String(String::from("src/main.rs")),
         );
@@ -253,12 +90,12 @@ fn rename_non_mutating_or_error_returns_failure(
 
     match scenario {
         FailureScenario::RelativeUri => {
-            let response = execute_request(&adapter, &request_with_args(rename_arguments))
+            let response = execute_request(&adapter, &request_with_args(arguments))
                 .expect("equivalent relative file URI should succeed");
             assert!(response.is_success());
         }
         FailureScenario::NoChange => {
-            let err = execute_request(&adapter, &request_with_args(rename_arguments))
+            let err = execute_request(&adapter, &request_with_args(arguments))
                 .expect_err("failure scenario should return Err");
             assert!(
                 err.message().contains("no content changes"),
@@ -267,7 +104,7 @@ fn rename_non_mutating_or_error_returns_failure(
             assert_eq!(err.reason_code(), Some(ReasonCode::SymbolNotFound));
         }
         FailureScenario::AdapterError => {
-            let err = execute_request(&adapter, &request_with_args(rename_arguments))
+            let err = execute_request(&adapter, &request_with_args(arguments))
                 .expect_err("failure scenario should return Err");
             assert!(
                 err.message().contains("rust-analyzer adapter failed"),
@@ -276,7 +113,7 @@ fn rename_non_mutating_or_error_returns_failure(
             assert_eq!(err.reason_code(), None);
         }
         FailureScenario::UriMismatch => {
-            let err = execute_request(&adapter, &request_with_args(rename_arguments))
+            let err = execute_request(&adapter, &request_with_args(arguments))
                 .expect_err("failure scenario should return Err");
             assert!(
                 err.message().contains("does not match file payload"),
@@ -285,7 +122,7 @@ fn rename_non_mutating_or_error_returns_failure(
             assert_eq!(err.reason_code(), Some(ReasonCode::IncompletePayload));
         }
         FailureScenario::InvalidUri => {
-            let err = execute_request(&adapter, &request_with_args(rename_arguments))
+            let err = execute_request(&adapter, &request_with_args(arguments))
                 .expect_err("failure scenario should return Err");
             assert!(
                 err.message()
@@ -295,112 +132,6 @@ fn rename_non_mutating_or_error_returns_failure(
             assert_eq!(err.reason_code(), Some(ReasonCode::IncompletePayload));
         }
     }
-}
-
-// ---------------------------------------------------------------------------
-// stdin/stdout dispatch layer tests (run_with_adapter)
-// ---------------------------------------------------------------------------
-
-fn valid_request_json() -> String {
-    let request = request_with_args(rename_arguments());
-    serde_json::to_string(&request).expect("serialize request")
-}
-
-/// Dispatches `input` through `run_with_adapter` and parses the response.
-fn dispatch_stdin(input: &[u8], adapter: &MockAdapter) -> PluginResponse {
-    let mut stdin = std::io::Cursor::new(input.to_vec());
-    let mut stdout = Vec::new();
-    run_with_adapter(&mut stdin, &mut stdout, adapter).expect("dispatch should succeed");
-    let output = String::from_utf8(stdout).expect("utf8 stdout");
-    serde_json::from_str(output.trim()).expect("parse response")
-}
-
-#[rstest]
-#[case::success(
-    format!("{}\n", valid_request_json()).into_bytes(),
-    adapter_returning(Ok(String::from("fn new_name() -> i32 {\n    1\n}\n"))),
-    true,
-    None
-)]
-#[case::empty_stdin(Vec::new(), adapter_unused(), false, Some("plugin request was empty"))]
-#[case::invalid_json(
-    b"not valid json\n".to_vec(),
-    adapter_unused(),
-    false,
-    Some("invalid plugin request JSON")
-)]
-fn run_with_adapter_dispatch_layer(
-    #[case] input: Vec<u8>,
-    #[case] adapter: MockAdapter,
-    #[case] expect_success: bool,
-    #[case] expected_message: Option<&str>,
-) {
-    let response = dispatch_stdin(&input, &adapter);
-    assert_eq!(response.is_success(), expect_success);
-
-    if let Some(needle) = expected_message {
-        assert!(
-            response
-                .diagnostics()
-                .iter()
-                .any(|diagnostic| diagnostic.severity() == DiagnosticSeverity::Error),
-            "expected at least one error diagnostic, got: {:?}",
-            response.diagnostics(),
-        );
-        assert!(
-            response
-                .diagnostics()
-                .iter()
-                .any(|diagnostic| diagnostic.message().contains(needle)),
-            "expected diagnostic mentioning '{needle}', got: {:?}",
-            response.diagnostics(),
-        );
-    }
-}
-
-#[rstest]
-#[case::missing_position(
-    {
-        let mut arguments = rename_arguments();
-        arguments.remove("position");
-        request_with_args(arguments)
-    },
-    ReasonCode::IncompletePayload
-)]
-#[case::unsupported_operation(
-    PluginRequest::new("extract_method", Vec::new()),
-    ReasonCode::OperationNotSupported
-)]
-fn failure_responses_include_reason_codes(
-    #[case] request: PluginRequest,
-    #[case] expected_reason: ReasonCode,
-) {
-    let input = format!(
-        "{}\n",
-        serde_json::to_string(&request).expect("serialize request")
-    );
-    let response = dispatch_stdin(input.as_bytes(), &adapter_unused());
-
-    assert!(!response.is_success());
-    assert!(
-        response
-            .diagnostics()
-            .iter()
-            .any(|diagnostic| diagnostic.reason_code() == Some(expected_reason)),
-        "expected reason code {expected_reason:?}, got: {:?}",
-        response.diagnostics(),
-    );
-}
-
-fn request_with_path(path: &str) -> PluginRequest {
-    PluginRequest::with_arguments(
-        "rename-symbol",
-        vec![FilePayload::new(
-            PathBuf::from(path),
-            "fn old_name() -> i32 {\n    1\n}\n",
-        )],
-        rename_arguments(),
-    )
 }
 
 #[rstest]
