+++ docs/roadmap.md
@@ -173,9 +173,12 @@ does* *not require source inspection or external runbooks.*
   - [x] Enable clap-provided `--version` and `-V` support.
   - [x] Add a `long_about` quick-start block aligned with the
         [user's guide](users-guide.md).
+  - [x] Validate the documentation updates with `make markdownlint` and
+        `make fmt` alongside the Rust quality gates.
   - [x] Acceptance criteria: `weaver --version` and `weaver -V` both exit 0
         and emit the same version string, and `weaver --help` includes at
-        least one runnable quick-start command example.
+        least one runnable quick-start command example, and `make check-fmt`,
+        `make markdownlint`, `make fmt`, `make lint`, and `make test` pass.
 - [ ] 2.2.4. Provide contextual guidance when a domain is supplied without an
       operation. See
       [Level 2](ui-gap-analysis.md#level-2--domain-without-operation-weaver-observe)
